requirements.txt
RPi.GPIO
Adafruit-Blinka
adafruit-circuitpython-busdevice
adafruit-circuitpython-ads1x15